# openapi.yaml
openapi: 3.0.0
info:
  title: Card Benefits Admin API
  description: |
    Admin management API for card and benefit administration.
    
    ## Rate Limiting
    All endpoints are rate limited per admin user:
    - List operations: 100-150 requests per minute
    - Create operations: 50 requests per minute
    - Update operations: 50 requests per minute
    - Delete operations: 30 requests per minute
    
    ## Pagination
    All list endpoints support pagination:
    - Default limit: 20-50 items per page (varies by endpoint)
    - Maximum limit: 100 items per page
    - Responses include pagination metadata (total, page, limit, totalPages, hasMore)
    
    ## Authentication
    All endpoints require admin role authentication via cookie (auth_token)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Development
...
paths:
...
  /api/admin/cards/{id}/benefits:
    get:
      tags:
        - Benefits
      summary: List benefits for card
      description: Get all benefits for a specific card
      operationId: listBenefits
      x-rateLimit:
        limit: 100
        window: minute
        description: 100 requests per minute per admin user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Benefits retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Benefit"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
      security:
        - cookieAuth: []

    post:
      tags:
        - Benefits
      summary: Create benefit for card
      description: Create a new benefit for a specific card
      operationId: createBenefit
      x-rateLimit:
        limit: 50
        window: minute
        description: 50 requests per minute per admin user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
                - stickerValue
                - resetCadence
              properties:
                name:
                  type: string
                  maxLength: 200
                type:
                  type: string
                  enum:
                    - INSURANCE
                    - CASHBACK
                    - TRAVEL
                    - BANKING
                    - POINTS
                    - OTHER
                stickerValue:
                  type: integer
                  minimum: 0
                resetCadence:
                  type: string
                  enum:
                    - ANNUAL
                    - QUARTERLY
                    - MONTHLY
                    - ONE_TIME
                isDefault:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Benefit created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Benefit"
      security:
        - cookieAuth: []
...
  /api/admin/audit-logs:
    get:
      tags:
        - Audit Logs
      summary: List audit logs
      operationId: listAuditLogs
      x-rateLimit:
        limit: 150
        window: minute
        description: 150 requests per minute per admin user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: actionType
          in: query
          schema:
            type: string
            enum:
              - CREATE
              - UPDATE
              - DELETE
        - name: resourceType
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Search (max 255 chars)
          schema:
            type: string
            maxLength: 255
      responses:
        "200":
          description: Audit logs retrieved
      security:
        - cookieAuth: []
...
components:
  schemas:
...
    Benefit:
      type: object
      properties:
        id:
          type: string
        masterCardId:
          type: string
        name:
          type: string
        type:
          type: string
        stickerValue:
          type: integer
        resetCadence:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HTTP cookie authentication
security:
  - cookieAuth: []